The Football Fest
The very famous football club Manchester United decided to popularize
football in India by organizing a football fest. The fest had many
events for different sections of people.
For the awesome coders of Hacker Earth, there was an event called PASS
and BACK. In this event, the coders were given
N
N passes and players having ids between
1
1 and
1000000
1000000.
Initially some player with a given id had the ball in his possession.
The coders had to make a program to display the id of the player who
possessed the ball after exactly
N
N passes.

Description of the passes:
There were two kinds of passes:
1.
P
P
I
D
ID
2.
B
B

Explanation :

for the first kind of pass, the player in possession of the ball passes
the ball to player with
i
d
=
I
D
id=ID while for the second kind of a pass, the player in possession of
the ball passes the ball back to the player who had passed the ball to
him.

NOTE:

It is guaranteed that the given order of all the passes will be a valid
order .

INPUT :

The first line of the input contains the number of test cases. For each
test case, two space separated integers
N
N and
I
D
ID ( of the player possessing the ball in the very beginning).
N
N lines follow describing the passes. ( for description of the passes,
refer the statement above. )

OUTPUT :

Output to each test case should be a single line containing the "Player"

I
D
ID (quotes for clarity) of the player who possesses the ball after
N
N passes.

CONSTRAINTS :

1
≤
T
≤
100
1≤T≤100
1
≤
N
≤
100000
1≤N≤100000
1
≤
I
D
≤
1000000